feat(auth): refactor Google sign-in for platform-specific interop
- Introduced platform-specific modules for Google sign-in:
  - `android.rs` for native Android integration via Firebase Auth.
  - `web.rs` for web-based Firebase Auth integration.
- Removed redundant Firebase imports in `index.js` and migrated logic to `web.js`.
- Updated `mod.rs` to conditionally include platform-specific modules.
- Commented out tmate setup in GitHub Actions workflow for debugging.
