+++ .github/workflows/pr-workflow.yml
@@ -318,28 +318,27 @@ jobs:
         run: |
           make test
 
-      - name: Setup SSH session with tmate
-        uses: mxschmitt/action-tmate@v3
-        if: failure()
-        timeout-minutes: 300
-        env:
-          DYNAMO_TABLE_PREFIX: ratel-local
-          ENV: local
-          RUSTFLAGS: "-D warnings"
-          RUST_LOG: info
-          DYNAMO_ENDPOINT: http://localstack:4566
-          AWS_ACCESS_KEY_ID: test
-          AWS_SECRET_ACCESS_KEY: test
-          AWS_REGION: ap-northeast-2
-          AWS_ACCOUNT_ID: test
-          COMMIT: pr-${{ github.sha }}
-          ECR: ratel/app-shell
-          PLAYWRIGHT_BASE_URL: http://localhost:8080
-          PLAYWRIGHT_TIMEOUT: 60000
+      # - name: Setup SSH session with tmate
+      #   uses: mxschmitt/action-tmate@v3
+      #   if: failure()
+      #   timeout-minutes: 300
+      #   env:
+      #     DYNAMO_TABLE_PREFIX: ratel-local
+      #     ENV: local
+      #     RUSTFLAGS: "-D warnings"
+      #     RUST_LOG: info
+      #     DYNAMO_ENDPOINT: http://localstack:4566
+      #     AWS_ACCESS_KEY_ID: test
+      #     AWS_SECRET_ACCESS_KEY: test
+      #     AWS_REGION: ap-northeast-2
+      #     AWS_ACCOUNT_ID: test
+      #     COMMIT: pr-${{ github.sha }}
+      #     ECR: ratel/app-shell
+      #     PLAYWRIGHT_BASE_URL: http://localhost:8080
+      #     PLAYWRIGHT_TIMEOUT: 60000
 
-        with:
-          limit-access-to-actor: true
-
+      #   with:
+      #     limit-access-to-actor: true
 
       - name: Move Playwright Tests
         if: always()
