+++ app/ratel/js/src/auth/index.js
@@ -1,8 +1,6 @@
-import * as firebase from "./firebase";
 import * as wallet from "./wallet";
 
 const auth = {
-  firebase,
   wallet,
 };
